+++ CHANGELOG.md
@@ -35,6 +35,7 @@ All notable changes to this project will be documented in this file.
 - [Core] More tests for Core (#4396 by: HebaruSan)
 - [Multiple] Exception handling revamp, parallel multi-host inflation (#4398 by: HebaruSan)
 - [Multiple] Make hash caches thread-safe, Netkan warning for uncompiled plugins (#4400 by: HebaruSan)
+- [Netkan] Make CachingHttpService more thread-safe (#4402 by: HebaruSan)
 
 ## v1.36.0 (Quasar)